活动页面 评分跳转bug

+++ src/main/resources/templates/client/activity_list.ftl
@@ -26,6 +26,21 @@ function unfinish()
 <#if alert?? &&alert == 1>
 window.onload=unfinish;
 </#if>
+
+function confirm()
+{
+var del=confirm("确认删除该活动？");
+if (del == true)
+  {
+  return true;
+  }
+else
+  {
+  return false;
+  }
+}
+
+
 </script>
 </head>
 <body>
@@ -92,7 +107,7 @@ window.onload=unfinish;
 		        		<td>${item.title!''}</td>
 		        		<td style="color:#0ab2cb;">${item.address!''}</td>
 		        		<td style="color:#e67817;">${item.activityType!''}</td>
-		        		<td><a href="/activity/edit?id=${item.id?c!''}">管理</a>丨<a href="/activity/check?id=${item.id?c!''}">查看</a>丨<a  href="/activity/delete?id=${item.id?c!''}">删除</a></td>
+		        		<td><a href="/activity/edit?id=${item.id?c!''}">管理</a>丨<a href="/activity/check?id=${item.id?c!''}">查看</a>丨<a  onclick="javascript:confirm();" href="/activity/delete?id=${item.id?c!''}">删除</a></td>
 		        	</tr>
 	        	</#list>
 	        </#if>
